Add files via upload
Files' name starting with "Z" is name of files which are totally created by me with no reference. I tried to use "Property" to write a script that is called "AbilitySystem". Firstly, attach two scripts on a game object, then while running the game, once the player click the mouse left button, the console would update the current ability information of players including strength, agile, speed, and luck. Meanwhile, the color of object would changed in order to visualize this progress.